Code finalised

+++ Assigment_1/GUI_1.m
@@ -575,7 +575,7 @@ function apply_LPF(hObject, eventdata, handles)
 		handles.idft = uint8(255*hsv2rgb(img_hsv));
 	else 
 		[M,N] = size(handles.dft_img);
-		handles.idft = handles.idft(1:M, 1:N);									% Carving out the original image back
+		handles.idft = uint8(handles.idft(1:M, 1:N));									% Carving out the original image back
 	end
 
 	% Enhancing the magnitude for display
@@ -628,7 +628,7 @@ function apply_HPF(hObject, eventdata, handles)
 		handles.idft = uint8(255*hsv2rgb(img_hsv));
 	else 
 		[M,N] = size(handles.dft_img);
-		handles.idft = handles.idft(1:M, 1:N);									% Carving out the original image back
+		handles.idft = uint8(handles.idft(1:M, 1:N));									% Carving out the original image back
 	end
 
 	% Enhancing the magnitude for display
